yolov4-deploy-model/seldom-server/yolov4/requirements.txt
seldon-core==1.6.0
numpy
tensorflow==2.4.0
Pillow